# Speed up EarthLocation.get_gcrs