Now that we have a faster method for calculating `obsgeoloc` and `obsgeovel` from an `EarthLocation`, there was no reason not to use it inside `EarthLocation` as well. The speed up is dramatic (see below). This partially addresses #11072.

Note that with this change, the new `include_velocity` argument introduced to `get_itrs()` is no longer used. I thought it made most sense to just revert it, as I think we still have larger decisions to make on how to deal with storing a location. 

It also changes the calculation slightly (see in-line comment); if it is felt this makes the code too unclear, I can revert that commit.

cc @mkbrewer

```
from astropy import units as u
from astropy.time import Time
from astropy.coordinates import EarthLocation
el = EarthLocation([10., 20.]*u.deg, [11, 21]*u.deg)
t = Time('J2010')
%timeit el.get_gcrs(t)
# current master: 33.9 ms ± 180 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
# first commit: 2.09 ms ± 11.5 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
# full PR: 1.74 ms ± 29.7 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```

EDIT: also
```
%timeit get_body('mars', t, location=el)
# current master: 44.3 ms ± 558 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
# this PR: 11.1 ms ± 147 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```