@mkbrewer 2020-11-22
This speed up is amazing! In that call to `get_body()`, are you using the JPL ephemeris or the built in one? If the later, I'd like to see the timing for the JPL ephemeris also as it has less overhead. 

I am very pleased that you have come around to adding this additional method to the `EarthLocation` class. This is exactly what I wished for in the first place. I would, however, request that you rename it `get_posvel()` and make it public. That way the user can easily retrieve the ground based observer's position and velocity in any of CIRS, TETE or TEME by simply calling this method with `gcrs_to_ref`set to an identity matrix as is done in the calls to `erfa.c2tcio()` in the relevant intermediate rotation transforms. In fact, you could make `gcrs_to_ref` optional with the identity matrix as the default. 

I don't think your calculation of the velocity in the target frame is unclear, in fact, I think it is quite innovative

With this change, we should revisit the intermediate rotation transforms for CIRS and TETE as this new method makes the helper function that you added there superfluous.