@mkbrewer 2020-11-22
One odd thing is that in `solarsystem.rst` there are two instances of retrieving the position of Jupiter. One using the built in ephemeris and the other using the JPL ephemeris. Both of these use the same location and time and yet the values of `obsgeopos` and `obsgeovel` are different. This PR's test using the built in ephemeris results in these values being closer to those shown in the JPL test.