@mkbrewer 2020-11-22
In `EarthLocation.get_posvel()` it might be easier for the user if the first argument was just `'cirs'`, `'tete'` or `'teme'`. Then if the second argument is a BPN rather than the default, this could be passed to `tete_to_itrs_mat()` in the case of `'tete'` to yield the GCRS position and velocity required by the TETE intermediate rotation transforms.