@mhvk 2020-11-22
@mkbrewer - I made a private method somewhat on purpose, as I think it helps separating getting the speed-up from how to expose a more flexible `get_posvel` to the user.

In an ideal world, I feel we should not need any of these methods as the transformation machinery itself would be fast enough; from that perspective, keeping everything in `intermediate_rotation_transformations` might actually make more sense than having a method on `EarthLocation` (indeed, I started on that yesterday and this morning had just decided to revert to that, but concluded again it was less good).

Anyway, let's do the speed-up first!