@mhvk 2020-11-22
The failure is very puzzling: I can understand a difference in `obsgeovel`, but where does this large difference in `obsgeoloc` come from? Investigating...