@mkbrewer 2020-11-22
Then in the TETE<->GCRS transforms, we could simply have: 
```
    loc_gcrs = tete_frame.location.get_posvel(tete_frame.obstime, 'tete', gcrs_to_ref=rbpn) 
```