@mhvk 2020-11-22
As you noticed, the JPL based example is much closer, even though of course `obsgeoloc` and `obsgeovel` do not depend on the ephemeris. Looking through the history, I see that that example was updated more recently, so something else must have (slightly) changed - maybe the IERS B tables?