@mkbrewer 2020-11-22
How about trying it using the JPL ephemeris then? If you do, please time it for both the master and this PR.