@mhvk 2020-11-22
OK, I updated the docs with the new values (which are within tolerance). As for timings with JPL:
```
solar_system_ephemeris.set('de432s')
# Same el and t as on top
%timeit get_body('mars', t, el)
# master  47.9 ms ± 214 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
# current PR 14.7 ms ± 529 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```
So, speed-up is very similar (since time was dominated by `get_gcrs_posvel`....)