@mkbrewer 2020-11-22
Hmm, I thought the JPL lookups would be quite a bit faster than the built in calculations. Perhaps, it's just the overhead associated with file I/O which would be mitigated for an array of times.

Since this results in a slight change in position, I'd appreciate it if you could run this script and post the output:

[skyfield_topo_test.py.txt](https://github.com/astropy/astropy/files/5579839/skyfield_topo_test.py.txt)

I haven't converted apparent position to use the TETE frame there, but you are welcome to do that.