@mhvk 2020-11-22
[astropy_topo_test_output.txt](https://github.com/astropy/astropy/files/5579868/astropy_topo_test_output.txt)