@mkbrewer 2020-11-22
Thanks. I'll compare this output to my program.