@mkbrewer 2020-11-22
For the most part the distance is good to 1 mm. There are just a few samples that are off by 1.5 - 2 mm.