@mkbrewer 2020-11-22
I'd need to add another decimal place to my output to see any real difference here.