@mhvk 2020-11-22
@mkbrewer - for a proper comparison, I wondered if one should transform to `TETE` rather than use `_apparent_position_in_true_coordinates` - as the latter infers `location` from `obsgeoloc` and `obsgeovel` rather than use the location directly, so may introduce small rounding errors. But I ran a quick check with a suitably modified version of your code and there are no differences at all in the numerical output.