@mkbrewer 2020-11-22
Yeah, I said that you were welcome to do that. I just wasn't real sure of myself on the proper syntax is all.

I profiled the conversion from GCRS to AltAz:
```
        1    0.000    0.000    0.364    0.364 intermediate_rotation_transforms.py:171(gcrs_to_cirs)
        1    0.000    0.000    0.096    0.096 intermediate_rotation_transforms.py:45(gcrs_to_cirs_mat)
        1    0.000    0.000    0.002    0.002 intermediate_rotation_transforms.py:50(cirs_to_itrs_mat)
        1    0.000    0.000    0.001    0.001 intermediate_rotation_transforms.py:96(get_location_gcrs)
        1    0.000    0.000    0.262    0.262 icrs_cirs_transforms.py:180(gcrs_to_gcrs)
```
Looks like most of the time was spent in `gcrs_to_gcrs()`. Since I don't see any GCRS<->ICRS calls, I guess this is just how long it took to decide that the location and times were equal.