@mhvk 2020-11-22
Here's the changes I made:
```
(test) mhvk@weasel:~$ diff astropy_topo_test.py~ astropy_topo_test.py
2,3c2
< from astropy.coordinates import solar_system_ephemeris, EarthLocation, get_body, get_body_barycentric, AltAz, SkyCoord
< from astropy.coordinates.solar_system import _apparent_position_in_true_coordinates
---
> from astropy.coordinates import solar_system_ephemeris, EarthLocation, get_body, get_body_barycentric, AltAz, SkyCoord, TETE
48c47,48
<         self.apparent = _apparent_position_in_true_coordinates(self.radec)
---
>         self.apparent = self.radec.transform_to(TETE(obstime=self.ast_times,
>                                                      location=self.loc))
```